rationai/visual/explain/requirements.txt
jupyterlab==2.2.8
pandas==1.1.3
tensorflow==2.3.1
scikit-learn==0.23.0
gitpython==3.1.9
Pillow==8.0.0
matplotlib==3.3.2
openslide-python==1.1.2
imgaug==0.4.0
keras==2.4.3
tf-keras-vis==0.5.3
seaborn==0.11.0
h5py==2.10
lime==0.2.0.1
